requirements.txt
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Django==5.1.7
django-apscheduler==0.7.0
djangorestframework==3.16.0
exceptiongroup==1.2.2
fastapi==0.115.11
h11==0.14.0
idna==3.10
packaging==24.2
psutil==7.0.0
pydantic==2.10.6
pydantic_core==2.27.2
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.1
requests==2.32.3
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.46.1
typing_extensions==4.12.2
tzlocal==5.3.1
urllib3==2.3.0
uvicorn==0.34.0
